+++ mc2/controllers/base/views.py
@@ -1,25 +1,27 @@
 import json
 import os.path
+import urllib
 from django.conf import settings
 from django.shortcuts import get_object_or_404, redirect
 from django.http import (
-    HttpResponse, HttpResponseNotFound, HttpResponseBadRequest)
+    HttpResponse, HttpResponseNotFound,
+    HttpResponseNotAllowed, HttpResponseServerError)
 from django.contrib.auth.decorators import (
     login_required, user_passes_test)
 from django.utils.decorators import method_decorator
 from django.views.decorators.csrf import csrf_exempt
 from django.views.generic.base import View
 from django.views.generic import TemplateView
-from django.views.generic.edit import UpdateView, CreateView
+from django.views.generic.edit import UpdateView, CreateView, FormMixin
 from django.core.urlresolvers import reverse, reverse_lazy
 from django.contrib import messages
 from mc2.organizations.utils import org_permission_required
 from mc2.organizations.utils import active_organization
+from mc2.organizations.models import Organization
 from mc2.controllers.base.models import Controller
 from mc2.controllers.base.forms import (
     ControllerFormHelper)
-from mc2.controllers.base import exceptions
-from mc2 import tasks
+from mc2.controllers.base import exceptions, tasks
 
 
 @login_required
@@ -39,7 +41,7 @@ def update_marathon_exists_json(request, controller_pk):
         content_type='application/json')
 
 
-class ControllerViewMixin(View):
+class ControllerViewMixin(FormMixin, View):
     pk_url_kwarg = 'controller_pk'
     permissions = []
     social_auth = None
@@ -72,6 +74,24 @@ class ControllerViewMixin(View):
             return Controller.objects.none()
         return Controller.objects.filter(organization=organization)
 
+    def get_form(self, *args, **kwargs):
+        """
+        Return a form with the organizations limited to the organizations
+        this user has access to. Returns a form with all availble
+        organizations for super users.
+
+        Since mixin is used in views that subclass CreateView or UpdateView
+        I am relying on those to provide the `get_form` as implemented
+        in the `FormMixin` both of those inherit
+        """
+        form = FormMixin.get_form(self, *args, **kwargs)
+        if self.request.user.is_superuser:
+            return form
+
+        form.controller_form.fields['organization'].queryset = (
+            self.request.user.organization_set.all())
+        return form
+
 
 class ControllerCreateView(ControllerViewMixin, CreateView):
     form_class = ControllerFormHelper
@@ -95,6 +115,48 @@ class ControllerCreateView(ControllerViewMixin, CreateView):
         return response
 
 
+class ControllerCloneView(ControllerViewMixin, CreateView):
+    form_class = ControllerFormHelper
+    template_name = 'controller_edit.html'
+    permissions = ['controllers.base.add_controller']
+
+    def get_initial(self):
+        initial = super(ControllerCloneView, self).get_initial()
+
+        controller = get_object_or_404(
+            Controller, pk=self.kwargs.get('controller_pk'))
+
+        initial.update({
+            'marathon_cpus': controller.marathon_cpus,
+            'marathon_mem': controller.marathon_mem,
+            'marathon_instances': controller.marathon_instances,
+            'marathon_cmd': controller.marathon_cmd,
+            'description': controller.description,
+            'envs': [
+                {'key': env.key, 'value': env.value}
+                for env in controller.env_variables.all()],
+            'labels': [
+                {'name': label.name, 'value': label.value}
+                for label in controller.label_variables.all()]})
+        return initial
+
+    def get_success_url(self):
+        return reverse("home")
+
+    def form_valid(self, form):
+        form.controller_form.instance.organization = active_organization(
+            self.request)
+        form.controller_form.instance.owner = self.request.user
+        form.controller_form.instance.save()
+
+        form.env_formset.instance = form.controller_form.instance
+        form.label_formset.instance = form.controller_form.instance
+
+        response = super(ControllerCloneView, self).form_valid(form)
+        tasks.start_new_controller.delay(form.controller_form.instance.id)
+        return response
+
+
 class ControllerEditView(ControllerViewMixin, UpdateView):
     form_class = ControllerFormHelper
     template_name = 'controller_edit.html'
@@ -103,16 +165,28 @@ class ControllerEditView(ControllerViewMixin, UpdateView):
     def get_queryset(self):
         return self.get_controllers_queryset(self.request)
 
+    def get_object(self):
+        controller = get_object_or_404(
+            Controller, pk=self.kwargs.get('controller_pk'))
+
+        if self.request.user.is_superuser:
+            return controller
+        elif controller.organization:
+            org = Organization.objects.for_user(self.request.user).filter(
+                pk=controller.organization.id).first()
+            if org and org.has_perms(
+                    self.request.user, ['base.change_controller']):
+                return controller
+        raise HttpResponseNotFound()
+
     def get_success_url(self):
         return reverse("home")
 
     def form_valid(self, form):
         response = super(ControllerEditView, self).form_valid(form)
-        try:
-            form.instance.update_marathon_app()
-        except exceptions.MarathonApiException:
-            messages.error(
-                self.request, 'Unable to update controller in marathon')
+        tasks.update_marathon_app.delay(form.instance.id)
+        messages.info(
+            self.request, '%s app update requested.' % form.instance.app_id)
         return response
 
 
@@ -129,18 +203,20 @@ class AppLogView(ControllerViewMixin, TemplateView):
             'controller': controller,
             'tasks': tasks,
             'task_ids': [t['id'].split('.', 1)[1] for t in tasks],
-            'scroll_backlog': (
-                self.request.GET.get('n') or settings.LOGDRIVER_BACKLOG)
+            'paths': ['stdout', 'stderr'],
+            'offset': self.request.GET.get('offset'),
+            'length': self.request.GET.get('length'),
         })
         return context
 
 
-class AppEventSourceView(ControllerViewMixin, View):
+class MesosFileLogView(ControllerViewMixin, View):
     def get(self, request, controller_pk, task_id, path):
         controller = get_object_or_404(
             self.get_controllers_queryset(request),
             pk=controller_pk)
-        n = request.GET.get('n') or settings.LOGDRIVER_BACKLOG
+        length = request.GET.get('length', '')
+        offset = request.GET.get('offset', '')
         if path not in ['stdout', 'stderr']:
             return HttpResponseNotFound('File not found.')
 
@@ -148,11 +224,23 @@ class AppEventSourceView(ControllerViewMixin, View):
         #       so as to not need to expose both in the templates.
         task = controller.infra_manager.get_controller_task_log_info(
             '%s.%s' % (controller.app_id, task_id))
+        file_path = os.path.join(task['task_dir'], path)
+
+        internal_redirect_url = settings.MESOS_FILE_API_PATH % {
+            'worker_host': task['task_host'],
+            'api_path': ('download.json'
+                         if request.GET.get('download')
+                         else 'read.json'),
+        }
 
         response = HttpResponse()
-        response['X-Accel-Redirect'] = '%s?n=%s' % (os.path.join(
-            settings.LOGDRIVER_PATH, task['task_host'],
-            task['task_dir'], path), n)
+        response['X-Accel-Redirect'] = '%s?%s' % (
+            internal_redirect_url, urllib.urlencode((
+                ('path', os.path.join(settings.MESOS_LOG_PATH, file_path)),
+                ('length', length),
+                ('offset', offset),
+            )))
+
         response['X-Accel-Buffering'] = 'no'
         return response
 
@@ -162,12 +250,9 @@ class ControllerRestartView(ControllerViewMixin, View):
 
     def get(self, request, controller_pk):
         controller = get_object_or_404(Controller, pk=controller_pk)
-        try:
-            controller.marathon_restart_app()
-            messages.info(self.request, 'App restart sent.')
-        except exceptions.MarathonApiException:
-            messages.error(
-                self.request, 'App restart failed. Please try again.')
+        tasks.marathon_restart_app.delay(controller.id)
+        messages.info(
+            self.request, '%s app restart requested.' % controller.app_id)
         return redirect('home')
 
 
@@ -182,13 +267,36 @@ class ControllerDeleteView(ControllerViewMixin, View):
 
     def post(self, request, controller_pk):
         controller = get_object_or_404(Controller, pk=controller_pk)
-        try:
-            controller.marathon_destroy_app()
-            controller.delete()
-            messages.info(self.request, 'App deletion sent.')
-        except exceptions.MarathonApiException:
-            msg = 'Failed to delete "%(id)s". Please try again.' % {
-                'id': controller.name}
-            messages.error(self.request, msg)
-            return HttpResponseBadRequest(msg)
+        tasks.marathon_destroy_app.delay(controller.id)
+        messages.info(
+            self.request, '%s app delete requested.' % controller.app_id)
         return redirect('home')
+
+
+class ControllerWebhookRestartView(View):
+    """
+    Unauthenticated webhook to restart an app.
+
+    Technically, this is a 'URL capability'.
+    """
+
+    @method_decorator(csrf_exempt)
+    def dispatch(self, *args, **kwargs):
+        return super(
+            ControllerWebhookRestartView, self).dispatch(*args, **kwargs)
+
+    def get(self, request, controller_pk, token):
+        return HttpResponseNotAllowed(['POST'])
+
+    def post(self, request, controller_pk, token):
+        controller = get_object_or_404(Controller, pk=controller_pk)
+        if str(controller.webhook_token) != token:
+            return HttpResponseNotFound()
+
+        try:
+            controller.marathon_restart_app()
+        except exceptions.MarathonApiException:
+            return HttpResponseServerError(
+                json.dumps({'error': 'Restart failed.'}),
+                content_type='application/json')
+        return HttpResponse(json.dumps({}), content_type='application/json')