+++ mc2/controllers/docker/tests/test_docker_controller.py
@@ -1,4 +1,5 @@
 import json
+import string
 
 import pytest
 import responses
@@ -13,7 +14,8 @@ from hypothesis.strategies import text, random_module, lists, just
 
 from mc2.controllers.base.models import EnvVariable, MarathonLabel
 from mc2.controllers.base.tests.base import ControllerBaseTestCase
-from mc2.controllers.docker.models import DockerController
+from mc2.controllers.docker.models import DockerController, traefik_domains
+from mc2.organizations.models import Organization
 
 
 def add_envvars(controller):
@@ -44,9 +46,17 @@ def docker_controller(with_envvars=True, with_labels=True, **kw):
     # so we remove them from the generated value.
     # TODO: Build a proper SlugField strategy.
     # TODO: Figure out why the field validation isn't being applied.
-    kw.setdefault("slug", text().map(
-        lambda t: "".join(t.replace(":", "").split())))
+    # Slugs must be domain-name friendly - used in the "generic" domain
+    slug = text(string.ascii_letters + string.digits + '-')
+    kw.setdefault("slug", slug)
+
     kw.setdefault("owner", models(User, is_active=just(True)))
+    kw.setdefault("organization", models(Organization, slug=slug))
+
+    # Prevent Hypothesis from generating domains with invalid characters
+    domain_urls = text(string.ascii_letters + string.digits + '-.')
+    kw.setdefault("domain_urls", domain_urls)
+
     # The model generator sees `controller_ptr` (from the PolymorphicModel
     # magic) as a mandatory field and objects if we don't provide a value for
     # it.
@@ -78,6 +88,7 @@ def check_and_clear_appdata(appdata, controller):
     check_and_remove_health(appdata, controller)
     check_and_remove_env(appdata, controller)
     check_and_remove_labels(appdata, controller)
+    check_and_remove_backoff_params(appdata)
     assert appdata == {}
 
 
@@ -110,19 +121,24 @@ def check_and_remove_health(appdata, controller):
     """
     Assert that the health check data is correct and remove it.
     """
-    if controller.marathon_health_check_path:
+    if controller.marathon_health_check_path and controller.port:
         assert appdata.pop("ports") == [0]
         assert appdata.pop("healthChecks") == [{
-            "gracePeriodSeconds": 3,
+            "gracePeriodSeconds": 60,
             "intervalSeconds": 10,
             "maxConsecutiveFailures": 3,
             "path": controller.marathon_health_check_path,
             "portIndex": 0,
             "protocol": "HTTP",
-            "timeoutSeconds": 5,
+            "timeoutSeconds": 20,
         }]
     assert "ports" not in appdata
     assert "healthChecks" not in appdata
+
+
+def check_and_remove_backoff_params(appdata):
+    appdata.pop('backoffFactor')
+    appdata.pop('backoffSeconds')
 
 
 def check_and_remove_env(appdata, controller):
@@ -146,6 +162,14 @@ def check_and_remove_labels(appdata, controller):
     domains = [u".".join([controller.app_id, settings.HUB_DOMAIN])]
     domains.extend(controller.domain_urls.split())
     assert sorted(labels.pop("domain").split()) == sorted(domains)
+    assert sorted(labels.pop("HAPROXY_0_VHOST").split()) == sorted(domains)
+    assert labels.pop("HAPROXY_GROUP") == "external"
+
+    traefik_domains = labels.pop("traefik.frontend.rule")
+    traefik_domains = traefik_domains.split(":", 2)[-1].split(",")
+    traefik_domains = [d.strip() for d in traefik_domains]
+    assert sorted(traefik_domains) == sorted(domains)
+
     # We may have duplicate keys in here, but hopefully the database always
     # return the objects in the same order.
     lvs = {lv.name: lv.value for lv in controller.label_variables.all()}
@@ -159,6 +183,34 @@ def check_and_remove_optional(appdata, field, value):
     if value:
         assert appdata.pop(field) == value
     assert field not in appdata
+
+
+def test_traefik_domains_single():
+    """
+    When a domains string with a single domain is passed to traefik_domains,
+    it returns a Host frontend rule.
+    """
+    domains = 'abc.com'
+    assert traefik_domains(domains) == 'Host: abc.com'
+
+
+def test_traefik_domains_multiple():
+    """
+    When a domains string with multiple domains is passed to traefik_domains,
+    it returns multiple Host frontend rules joined by ';'.
+    """
+    domains = 'abc.com def.co.za   ghi.co.ng'
+    assert (traefik_domains(domains) ==
+            'Host: abc.com, def.co.za, ghi.co.ng')
+
+
+def test_traefik_domains_none():
+    """
+    When a domains string with no domains is passed to traefik_domains, it
+    returns an empty string.
+    """
+    domains = '  '
+    assert traefik_domains(domains) == ''
 
 
 @pytest.mark.django_db
@@ -189,7 +241,7 @@ class DockerControllerHypothesisTestCase(TestCase):
         """
         app_data = controller.get_marathon_app_data()
         new_controller = DockerController.from_marathon_app_data(
-            controller.owner, app_data)
+            controller.owner, controller.organization, app_data)
         assert app_data == new_controller.get_marathon_app_data()
 
     @hsettings(perform_health_check=False, max_examples=50)
@@ -204,7 +256,7 @@ class DockerControllerHypothesisTestCase(TestCase):
         """
         app_data = controller.get_marathon_app_data()
         new_controller = DockerController.from_marathon_app_data(
-            controller.owner, app_data, name=name)
+            controller.owner, controller.organization, app_data, name=name)
         assert new_controller.name == name
 
         app_data_with_name = json.loads(json.dumps(app_data))
@@ -227,9 +279,15 @@ class DockerControllerHypothesisTestCase(TestCase):
         user.save()
         client = Client()
         assert client.login(username=user.username, password="password")
-        resp = client.post(
-            reverse('controllers.docker:hidden_import'),
-            {"name": name, "app_data": json.dumps(app_data)})
+        with responses.RequestsMock() as rsps:
+            # If the Marathon request hasn't occurred by the time this context
+            # is closed then the test will fail.
+            rsps.add(
+                responses.POST, '%s/v2/apps' % settings.MESOS_MARATHON_HOST,
+                body="{}", content_type="application/json", status=201)
+            resp = client.post(
+                reverse('controllers.docker:hidden_import'),
+                {"name": name, "app_data": json.dumps(app_data)})
         assert resp.status_code == 302
 
         new_controller = DockerController.objects.exclude(
@@ -259,16 +317,21 @@ class DockerControllerTestCase(ControllerBaseTestCase):
 
         custom_urls = "testing.com url.com"
         controller.domain_urls += custom_urls
+        domain_label = "{}.{} {}".format(
+            controller.app_id, settings.HUB_DOMAIN, custom_urls)
         self.assertEquals(controller.get_marathon_app_data(), {
             "id": controller.app_id,
             "cpus": 0.1,
             "mem": 128.0,
             "instances": 1,
             "cmd": "ping",
+            "backoffFactor": settings.MESOS_DEFAULT_BACKOFF_FACTOR,
+            "backoffSeconds": settings.MESOS_DEFAULT_BACKOFF_SECONDS,
             "labels": {
-                "domain": "{}.{} {}".format(controller.app_id,
-                                            settings.HUB_DOMAIN,
-                                            custom_urls),
+                "domain": domain_label,
+                "HAPROXY_GROUP": "external",
+                "HAPROXY_0_VHOST": domain_label,
+                "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
             },
             "container": {
@@ -284,16 +347,21 @@ class DockerControllerTestCase(ControllerBaseTestCase):
         controller.port = 1234
         controller.save()
 
+        domain_label = "{}.{} {}".format(
+            controller.app_id, settings.HUB_DOMAIN, custom_urls)
         self.assertEquals(controller.get_marathon_app_data(), {
             "id": controller.app_id,
             "cpus": 0.1,
             "mem": 128.0,
             "instances": 1,
             "cmd": "ping",
+            "backoffFactor": settings.MESOS_DEFAULT_BACKOFF_FACTOR,
+            "backoffSeconds": settings.MESOS_DEFAULT_BACKOFF_SECONDS,
             "labels": {
-                "domain": "{}.{} {}".format(controller.app_id,
-                                            settings.HUB_DOMAIN,
-                                            custom_urls),
+                "domain": domain_label,
+                "HAPROXY_GROUP": "external",
+                "HAPROXY_0_VHOST": domain_label,
+                "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App"
             },
             "container": {
@@ -310,16 +378,21 @@ class DockerControllerTestCase(ControllerBaseTestCase):
         controller.marathon_health_check_path = '/health/path/'
         controller.save()
 
+        domain_label = "{}.{} {}".format(
+            controller.app_id, settings.HUB_DOMAIN, custom_urls)
         self.assertEquals(controller.get_marathon_app_data(), {
             "id": controller.app_id,
             "cpus": 0.1,
             "mem": 128.0,
             "instances": 1,
             "cmd": "ping",
+            "backoffFactor": settings.MESOS_DEFAULT_BACKOFF_FACTOR,
+            "backoffSeconds": settings.MESOS_DEFAULT_BACKOFF_SECONDS,
             "labels": {
-                "domain": "{}.{} {}".format(controller.app_id,
-                                            settings.HUB_DOMAIN,
-                                            custom_urls),
+                "domain": domain_label,
+                "HAPROXY_GROUP": "external",
+                "HAPROXY_0_VHOST": domain_label,
+                "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
             },
             "container": {
@@ -333,13 +406,13 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             },
             "ports": [0],
             "healthChecks": [{
-                "gracePeriodSeconds": 3,
+                "gracePeriodSeconds": 60,
                 "intervalSeconds": 10,
                 "maxConsecutiveFailures": 3,
                 "path": '/health/path/',
                 "portIndex": 0,
                 "protocol": "HTTP",
-                "timeoutSeconds": 5
+                "timeoutSeconds": 20
             }]
         })
 
@@ -347,16 +420,21 @@ class DockerControllerTestCase(ControllerBaseTestCase):
         controller.volume_path = "/deploy/media/"
         controller.save()
 
+        domain_label = "{}.{} {}".format(
+            controller.app_id, settings.HUB_DOMAIN, custom_urls)
         self.assertEquals(controller.get_marathon_app_data(), {
             "id": controller.app_id,
             "cpus": 0.1,
             "mem": 128.0,
             "instances": 1,
             "cmd": "ping",
+            "backoffFactor": settings.MESOS_DEFAULT_BACKOFF_FACTOR,
+            "backoffSeconds": settings.MESOS_DEFAULT_BACKOFF_SECONDS,
             "labels": {
-                "domain": "{}.{} {}".format(controller.app_id,
-                                            settings.HUB_DOMAIN,
-                                            custom_urls),
+                "domain": domain_label,
+                "HAPROXY_GROUP": "external",
+                "HAPROXY_0_VHOST": domain_label,
+                "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
             },
             "container": {
@@ -377,29 +455,34 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             },
             "ports": [0],
             "healthChecks": [{
-                "gracePeriodSeconds": 3,
+                "gracePeriodSeconds": 60,
                 "intervalSeconds": 10,
                 "maxConsecutiveFailures": 3,
                 "path": '/health/path/',
                 "portIndex": 0,
                 "protocol": "HTTP",
-                "timeoutSeconds": 5
+                "timeoutSeconds": 20
             }]
         })
 
         controller.volume_path = ""
         controller.save()
 
+        domain_label = "{}.{} {}".format(
+            controller.app_id, settings.HUB_DOMAIN, custom_urls)
         self.assertEquals(controller.get_marathon_app_data(), {
             "id": controller.app_id,
             "cpus": 0.1,
             "mem": 128.0,
             "instances": 1,
             "cmd": "ping",
+            "backoffFactor": settings.MESOS_DEFAULT_BACKOFF_FACTOR,
+            "backoffSeconds": settings.MESOS_DEFAULT_BACKOFF_SECONDS,
             "labels": {
-                "domain": "{}.{} {}".format(controller.app_id,
-                                            settings.HUB_DOMAIN,
-                                            custom_urls),
+                "domain": domain_label,
+                "HAPROXY_GROUP": "external",
+                "HAPROXY_0_VHOST": domain_label,
+                "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
             },
             "container": {
@@ -422,13 +505,13 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             },
             "ports": [0],
             "healthChecks": [{
-                "gracePeriodSeconds": 3,
+                "gracePeriodSeconds": 60,
                 "intervalSeconds": 10,
                 "maxConsecutiveFailures": 3,
                 "path": '/health/path/',
                 "portIndex": 0,
                 "protocol": "HTTP",
-                "timeoutSeconds": 5
+                "timeoutSeconds": 20
             }]
         })
 
@@ -441,16 +524,21 @@ class DockerControllerTestCase(ControllerBaseTestCase):
         )
         self.mk_env_variable(controller)
 
+        domain_label = "{}.{}".format(controller.app_id, settings.HUB_DOMAIN)
         self.assertEquals(controller.get_marathon_app_data(), {
             "id": controller.app_id,
             "cpus": 0.1,
             "mem": 128.0,
             "instances": 1,
             "cmd": "ping",
+            "backoffFactor": settings.MESOS_DEFAULT_BACKOFF_FACTOR,
+            "backoffSeconds": settings.MESOS_DEFAULT_BACKOFF_SECONDS,
             "env": {"TEST_KEY": "a test value"},
             "labels": {
-                "domain": "{}.{}".format(controller.app_id,
-                                         settings.HUB_DOMAIN),
+                "domain": domain_label,
+                "HAPROXY_GROUP": "external",
+                "HAPROXY_0_VHOST": domain_label,
+                "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
             },
             "container": {
@@ -473,16 +561,21 @@ class DockerControllerTestCase(ControllerBaseTestCase):
         self.mk_env_variable(controller)
         self.mk_labels_variable(controller)
 
+        domain_label = "{}.{}".format(controller.app_id, settings.HUB_DOMAIN)
         self.assertEquals(controller.get_marathon_app_data(), {
             "id": controller.app_id,
             "cpus": 0.1,
             "mem": 128.0,
             "instances": 1,
             "cmd": "ping",
+            "backoffFactor": settings.MESOS_DEFAULT_BACKOFF_FACTOR,
+            "backoffSeconds": settings.MESOS_DEFAULT_BACKOFF_SECONDS,
             "env": {"TEST_KEY": "a test value"},
             "labels": {
-                "domain": "{}.{}".format(controller.app_id,
-                                         settings.HUB_DOMAIN),
+                "domain": domain_label,
+                "HAPROXY_GROUP": "external",
+                "HAPROXY_0_VHOST": domain_label,
+                "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
                 "TEST_LABELS_NAME": 'a test label value'
             },
@@ -525,14 +618,19 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             docker_image='docker/image',
         )
 
+        domain_label = "{}.{}".format(controller.app_id, settings.HUB_DOMAIN)
         self.assertEquals(controller.get_marathon_app_data(), {
             "id": controller.app_id,
             "cpus": 0.1,
             "mem": 128.0,
             "instances": 1,
+            "backoffFactor": settings.MESOS_DEFAULT_BACKOFF_FACTOR,
+            "backoffSeconds": settings.MESOS_DEFAULT_BACKOFF_SECONDS,
             "labels": {
-                "domain": "{}.{}".format(controller.app_id,
-                                         settings.HUB_DOMAIN),
+                "domain": domain_label,
+                "HAPROXY_GROUP": "external",
+                "HAPROXY_0_VHOST": domain_label,
+                "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
             },
             "container": {
@@ -544,3 +642,59 @@ class DockerControllerTestCase(ControllerBaseTestCase):
                 }
             }
         })
+
+    @responses.activate
+    def test_get_marathon_app_data_using_health_timeout_strings(self):
+        controller = DockerController.objects.create(
+            name='Test App',
+            owner=self.user,
+            marathon_cmd='ping',
+            docker_image='docker/image',
+            marathon_health_check_path='/health/path/',
+            port=1234,
+        )
+
+        custom_urls = "testing.com url.com"
+        controller.domain_urls += custom_urls
+        with self.settings(
+            MESOS_DEFAULT_GRACE_PERIOD_SECONDS='600',
+            MESOS_DEFAULT_INTERVAL_SECONDS='100',
+                MESOS_DEFAULT_TIMEOUT_SECONDS='200'):
+            domain_label = "{}.{} {}".format(
+                controller.app_id, settings.HUB_DOMAIN, custom_urls)
+            self.assertEquals(controller.get_marathon_app_data(), {
+                "id": controller.app_id,
+                "cpus": 0.1,
+                "mem": 128.0,
+                "instances": 1,
+                "cmd": "ping",
+                "backoffFactor": settings.MESOS_DEFAULT_BACKOFF_FACTOR,
+                "backoffSeconds": settings.MESOS_DEFAULT_BACKOFF_SECONDS,
+                "labels": {
+                    "domain": domain_label,
+                    "HAPROXY_GROUP": "external",
+                    "HAPROXY_0_VHOST": domain_label,
+                    "traefik.frontend.rule": traefik_domains(domain_label),
+                    "name": "Test App",
+                },
+                "container": {
+                    "type": "DOCKER",
+                    "docker": {
+                        "image": "docker/image",
+                        "forcePullImage": True,
+                        "network": "BRIDGE",
+                        "portMappings": [
+                            {"containerPort": 1234, "hostPort": 0}],
+                    }
+                },
+                "ports": [0],
+                "healthChecks": [{
+                    "gracePeriodSeconds": 600,
+                    "intervalSeconds": 100,
+                    "maxConsecutiveFailures": 3,
+                    "path": '/health/path/',
+                    "portIndex": 0,
+                    "protocol": "HTTP",
+                    "timeoutSeconds": 200
+                }]
+            })